Improved the style of the menus package
It still might not be perfect so feel free to update, but it's much
closer to PEP8 and PEP257 now. Also, don't bother PEP8:ing my >80
character lines unless you find a meaningful way to do so (randomly
hitting enter in an awkward position is not meaningful nor useful to
anyone).