Updated the loading player's data for the first time
Players' data now gets loaded from a database only when they join the
server or when the plugin is loaded. If someone manages to sneak onto
the server without using player_activate or showing up in PlayerIter,
that's not our fault and should be reported to SP or Valve. Hopefully
this makes everything a bit more clear on the code side!