Fixed a problem with item purchasing
Every time you purchased an item the item_purchased method was being
called. So now have changed to only call the single method rather than
all of the items owned. However this means we can't have specific
reactions to items for heroes. So I will fix it properly later; as this
is just a work around atm.